refactored race
