+++ src/drawers/race_drawer.py
@@ -1,5 +1,7 @@
 import pygame
 from tools import rotate_point
+from environment import Environment
+import pandas as pd
 
 
 class RaceDrawer:
@@ -19,10 +21,22 @@ class RaceDrawer:
     BOAT_COLOR = (255, 255, 255)
     BOAT_SCALE = 0.1
 
-    def __init__(self, screen):
-        self._screen = screen
+    SIZE = 1024, 768
+    BG_COLOR = 0, 0, 0
+    TEXT_COLOR = 255, 255, 255
+
+    def __init__(self, strategies: list, env: Environment):
+        self._strategies = strategies
+        self._env = env
         self._offset = (0, 0)
         self._scale = 0
+
+        pygame.init()
+        pygame.font.init()
+
+        self._font = pygame.font.SysFont('Arial', 30)
+        self._smallfont = pygame.font.SysFont('Arial', 20)
+        self._screen = pygame.display.set_mode(self.SIZE)
 
     def autoscale(self, buoys):
         """ Scale race canvas to fit all buoys """
@@ -98,4 +112,39 @@ class RaceDrawer:
             x, y = self.translate_pos(position)
             pygame.draw.circle(self._screen, self.BUOY_COLOR, (x, y), 5)
 
+    def write_text(self, text, row, color=(255, 255, 255)):
+        pos = 740, 30 + (row * 30)
+        textsurface = self._font.render(text, True, color)
+        self._screen.blit(textsurface, pos)
 
+    def draw(self):
+        self._screen.fill(self.BG_COLOR)
+        self.draw_env(self._env)
+        self.draw_buoys(self._env.get_buoys())
+
+        # draw all boats
+        scoreboard = []
+        for i, strategy in enumerate(self._strategies):
+
+            # update boat and draw
+            boat = strategy.get_boat()
+            self.draw_boat(boat)
+
+            # update scoreboard
+            scoreboard.append({
+                'name': strategy.get_name(),
+                'color': boat.get_boat_color(),
+                'marks_passed': boat.get_marks_passed(),
+                'dtw': boat.get_distance_to_waypoint()
+            })
+
+        # show scoreboard
+        scoreboard = pd.DataFrame(scoreboard).sort_values(by=['marks_passed', 'dtw'], ascending=[False, True])
+        i = 0
+        for _, row in scoreboard.iterrows():
+            text = "%s (DTW: %dm)" % (row['name'], row.dtw)
+            self.write_text(text, i, row.color)
+            i += 1
+
+        # display new frame
+        pygame.display.flip()
