+++ src/simulators/race_simulator.py
@@ -35,27 +35,12 @@ class RaceUpdateThread (threading.Thread):
 
 
 class RaceSimulator:
-    SIZE = 1024, 768
-    BG_COLOR = 0, 0, 0
-    TEXT_COLOR = 255, 255, 255
 
     def __init__(self, env, strategies):
         self._env = env
         self._strategies = strategies
-
-        pygame.init()
-        pygame.font.init()
-
-        self._font = pygame.font.SysFont('Arial', 30)
-        self._smallfont = pygame.font.SysFont('Arial', 20)
-        self._screen = pygame.display.set_mode(self.SIZE)
-        self._drawer = RaceDrawer(self._screen)
+        self._drawer = RaceDrawer(strategies, env)
         self._clock = pygame.time.Clock()
-
-    def write_text(self, text, row, color=(255, 255, 255)):
-        pos = 740, 30 + (row * 30)
-        textsurface = self._font.render(text, True, color)
-        self._screen.blit(textsurface, pos)
 
     def run(self):
         # scale the race canvas
@@ -67,44 +52,20 @@ class RaceSimulator:
         thread.start()
 
         while 1:
-            self._screen.fill(self.BG_COLOR)
             self._env.update()
 
-            self._drawer.draw_env(self._env)
-            self._drawer.draw_buoys(self._env.get_buoys())
-
             # update all boats
-            scoreboard = []
-            for i, strategy in enumerate(self._strategies):
-
-                # update boat and draw
+            for strategy in self._strategies:
                 boat = strategy.get_boat()
                 boat.update()
-                self._drawer.draw_boat(boat)
-
-                # update scoreboard
-                scoreboard.append({
-                    'name': strategy.get_name(),
-                    'color': boat.get_boat_color(),
-                    'marks_passed': boat.get_marks_passed(),
-                    'dtw': boat.get_distance_to_waypoint()
-                })
 
                 # update boat with current fps
                 fps = self._clock.get_fps()
                 if fps > 0:
                     boat.set_draw_fps(fps)
 
-            # show scoreboard
-            scoreboard = pd.DataFrame(scoreboard).sort_values(by=['marks_passed', 'dtw'], ascending=[False, True])
-            i = 0
-            for _, row in scoreboard.iterrows():
-                text = "%s (DTW: %dm)" % (row['name'], row.dtw)
-                self.write_text(text, i, row.color)
-                i += 1
-
-            # display new frame
-            pygame.display.flip()
+            # draw objects
+            self._drawer.draw()
 
             # check key events
             for event in pygame.event.get():
